requirements.txt
pandas>=1.5
numpy>=1.24
yfinance>=0.2.22
pandas_ta>=0.4.67b0    # maintained version compatible with Python 3.12
matplotlib>=3.7
streamlit>=1.24
pyyaml
backtesting>=0.3.0     # optional, for quick backtesting
requests
python-dotenv

# Alpaca Trading API
alpaca-trade-api>=3.2.0
